refactor(services): consolidate data formatting into generic utility
Move certificate, project and skill specific formatters into a single generic formatter utility.
Update services and components to use the new formatter and standardized types.
